Remove file plugin.hpp which should not be included on its own.
