+++ include/appbase/application_instance.hpp
@@ -4,6 +4,7 @@ namespace appbase {
 
    using executor_t = appbase_executor;
 
+   // ------------------------------------------------------------------------------------------
    class application : public application_base {
    public:
       static application&  instance() {
@@ -49,12 +50,62 @@ namespace appbase {
       inline static std::unique_ptr<application> app_instance;
       executor_t executor_;
    };
-}
 
-#include <appbase/plugin.hpp>
+   // ------------------------------------------------------------------------------------------
+   template<typename Impl>
+   class plugin : public abstract_plugin {
+      public:
+         plugin():_name(boost::core::demangle(typeid(Impl).name())){}
+         virtual ~plugin(){}
 
-namespace appbase {
+         virtual state get_state()const final         { return _state; }
+         virtual const std::string& name()const final { return _name; }
 
+         virtual void register_dependencies() {
+            static_cast<Impl*>(this)->plugin_requires([&](auto& plug){});
+         }
+
+         virtual void initialize(const variables_map& options) final {
+            if(_state == registered) {
+               _state = initialized;
+               static_cast<Impl*>(this)->plugin_requires([&](auto& plug){ plug.initialize(options); });
+               static_cast<Impl*>(this)->plugin_initialize(options);
+               //ilog( "initializing plugin ${name}", ("name",name()) );
+               app().plugin_initialized(*this);
+            }
+            assert(_state == initialized); /// if initial state was not registered, final state cannot be initialized
+         }
+
+         virtual void handle_sighup() override {
+         }
+
+         virtual void startup() final {
+            if(_state == initialized) {
+               _state = started;
+               static_cast<Impl*>(this)->plugin_requires([&](auto& plug){ plug.startup(); });
+               static_cast<Impl*>(this)->plugin_startup();
+               app().plugin_started(*this);
+            }
+            assert(_state == started); // if initial state was not initialized, final state cannot be started
+         }
+
+         virtual void shutdown() final {
+            if(_state == started) {
+               _state = stopped;
+               //ilog( "shutting down plugin ${name}", ("name",name()) );
+               static_cast<Impl*>(this)->plugin_shutdown();
+            }
+         }
+
+      protected:
+         plugin(const string& name) : _name(name){}
+
+      private:
+         state _state = abstract_plugin::registered;
+         std::string _name;
+   };
+
+   // ------------------------------------------------------------------------------------------
    template<typename Data, typename DispatchPolicy>
    void channel<Data,DispatchPolicy>::publish(int priority, const Data& data) {
       if (has_subscribers()) {
@@ -65,6 +116,7 @@ namespace appbase {
       }
    }
 
+   // ------------------------------------------------------------------------------------------
    class scoped_app {
    public:
       explicit scoped_app()  { assert(application::null_app_singleton()); app_ = &app(); }
