--- include/appbase/plugin.hpp
@@ -1,67 +1,0 @@
-#pragma once
-
-// we can remove this forward decl. later when we update tests/trx_generator/trx_generator.hpp
-// to include application.hpp instead of plugin.hpp.
-namespace appbase {
-   class application;
-   static application& app();
-}
-
-#include <appbase/abstract_plugin.hpp>
-
-namespace appbase {
-
-   template<typename Impl>
-   class plugin : public abstract_plugin {
-      public:
-         plugin():_name(boost::core::demangle(typeid(Impl).name())){}
-         virtual ~plugin(){}
-
-         virtual state get_state()const final         { return _state; }
-         virtual const std::string& name()const final { return _name; }
-
-         virtual void register_dependencies() {
-            static_cast<Impl*>(this)->plugin_requires([&](auto& plug){});
-         }
-
-         virtual void initialize(const variables_map& options) final {
-            if(_state == registered) {
-               _state = initialized;
-               static_cast<Impl*>(this)->plugin_requires([&](auto& plug){ plug.initialize(options); });
-               static_cast<Impl*>(this)->plugin_initialize(options);
-               //ilog( "initializing plugin ${name}", ("name",name()) );
-               app().plugin_initialized(*this);
-            }
-            assert(_state == initialized); /// if initial state was not registered, final state cannot be initialized
-         }
-
-         virtual void handle_sighup() override {
-         }
-
-         virtual void startup() final {
-            if(_state == initialized) {
-               _state = started;
-               static_cast<Impl*>(this)->plugin_requires([&](auto& plug){ plug.startup(); });
-               static_cast<Impl*>(this)->plugin_startup();
-               app().plugin_started(*this);
-            }
-            assert(_state == started); // if initial state was not initialized, final state cannot be started
-         }
-
-         virtual void shutdown() final {
-            if(_state == started) {
-               _state = stopped;
-               //ilog( "shutting down plugin ${name}", ("name",name()) );
-               static_cast<Impl*>(this)->plugin_shutdown();
-            }
-         }
-
-      protected:
-         plugin(const string& name) : _name(name){}
-
-      private:
-         state _state = abstract_plugin::registered;
-         std::string _name;
-   };
-
-}
